Create Glob Tutorial
This was written to solve an issue where previous analysts were using excel to manually copy and paste files together or sometimes using MS Access. This still required a lot of cleaning, making it a time consuming task. 

Most of the files were straight forward and located in the same directory folder for each fiscal year. However, 2018 had two issues. One was a series of files with "-merged.xlsx" added to the end of the files names for 2 of the years which were being pulled. I used the longer statements to filter out the "-merged.xlsx" files and began with either two digits and a letter (a was non-merged, b was merged). The second was that not all the files were saved in this filename format, so I wrote two different statements and appended the data frames in succession to get around this issue.

At the end, all the data frames were appended to each other to create a master file of all the training data for individuals in the system.